requirements.txt
mcstatus
aiofiles
asyncpg
tortoise-orm
python-dotenv